Authentication chain limited to publishing handler
